requirements.txt
aerich==0.5.8
aiosqlite==0.16.1
asgiref==3.4.1
asyncpg==0.24.0
bcrypt==3.2.0
cffi==1.14.6
click==8.0.1
cryptography==3.4.8
ddlparse==1.10.0
dictdiffer==0.9.0
ecdsa==0.17.0
fastapi==0.68.1
fastapi-utils==0.2.1
greenlet==1.1.1
h11==0.12.0
httptools==0.2.0
iso8601==0.1.16
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pyparsing==2.4.7
pypika-tortoise==0.1.1
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.23
starlette==0.14.2
tortoise-orm==0.17.7
typing-extensions==3.10.0.2
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==9.1
